// Client/src/app/shared/student.model.ts
export interface student {
    Password: any;
    FirstName: any;
    LastName: any;
   
    
    StudentID : number;

   
    EmailAddress : string;
    IdNumber : string;
    ContactNumber: string;
    Address: string;
    PreferedMethod: string;
    ProfileImage: string;
    

    StudentEnrolementID : number;
    Institution : string;
    Qualification: string;
    QualificationType: string;
    GraduationDate : string;
    DateRegistered : string;
    AverageToDate : string;



    NAME: number;
    Surname : string;
    Email : string;
    Phone : string;
    Gender : string ;
    ClientID: number;
    description:string;
    id: number;
    IsAccepted:number;  
   
   
    
  
  

    

      
}